Add support for JUNG and JGraphT to Java library
This commit adds the capability to covert FNSS Topology objects into
JUNG and JGraphT graph objects. JUNG and JGraphT are the most used
network graph libraries for Java. Converting FNSS topologies into
JUNG or JGraphT objects enables to use their implementations
of graph algorithms, such as Dijkstra algorithms and many more.

This commit adds the source of the classes implementing this
functionality as well as examples, test code and Jar files of the
new dependencies.

As a result of the implementation of this new feature, the FNSS Java
library depends on more third-party Java libraries: JUNG and JGraphT
as well as their dependencies.